+++ src/Http/Controllers/TimoneiroBaseController.php
@@ -14,6 +14,8 @@ class TimoneiroBaseController extends Controller
     /**
      * @param Request $request
      *
+     * @throws \Illuminate\Validation\ValidationException
+     *
      * @return Factory|\Illuminate\View\View
      */
     public function index(Request $request)
@@ -76,6 +78,14 @@ class TimoneiroBaseController extends Controller
         return Timoneiro::view($view, $viewData);
     }
 
+    /**
+     * @param Request $request
+     * @param $id
+     *
+     * @throws \Illuminate\Validation\ValidationException
+     *
+     * @return \Illuminate\View\View
+     */
     public function edit(Request $request, $id)
     {
         $dataType = $request->getDataType();
@@ -104,6 +114,8 @@ class TimoneiroBaseController extends Controller
      * @param Request $request
      * @param $id
      *
+     * @throws \Illuminate\Validation\ValidationException
+     *
      * @return \Illuminate\Http\RedirectResponse
      */
     public function update(Request $request, $id)
@@ -120,9 +132,18 @@ class TimoneiroBaseController extends Controller
         $request->check('edit', $data);
         $this->getService($dataType)->update($data, $request->all());
 
+        Timoneiro::pushNotification("Successfully Updated $dataType->display_name_singular");
+
         return redirect()->route("timoneiro.{$dataType->slug}.index");
     }
 
+    /**
+     * @param Request $request
+     *
+     * @throws \Illuminate\Validation\ValidationException
+     *
+     * @return \Illuminate\View\View
+     */
     public function create(Request $request)
     {
         $dataType = $request->getDataType();
@@ -143,6 +164,8 @@ class TimoneiroBaseController extends Controller
     /**
      * @param Request $request
      *
+     * @throws \Illuminate\Validation\ValidationException
+     *
      * @return \Illuminate\Http\RedirectResponse
      */
     public function store(Request $request)
@@ -151,6 +174,41 @@ class TimoneiroBaseController extends Controller
 
         $request->check('add');
         $this->getService($dataType)->create($request->all());
+
+        Timoneiro::pushNotification("Successfully Created $dataType->display_name_singular");
+
+        return redirect()->route("timoneiro.{$dataType->slug}.index");
+    }
+
+    /**
+     * @param Request $request
+     * @param $id
+     *
+     * @throws \Illuminate\Validation\ValidationException
+     *
+     * @return \Illuminate\Http\RedirectResponse
+     */
+    public function destroy(Request $request, $id)
+    {
+        $dataType = $request->getDataType();
+        $service = $this->getService($dataType);
+        $request->check('delete', $service->getModel());
+        $ids = [];
+
+        $displayName = count($ids) > 1 ? $dataType->display_name_plural : $dataType->display_name_singular;
+
+        if (empty($id)) {
+            $ids = explode(',', $request->ids);
+        } else {
+            $ids[] = $id;
+        }
+
+        try {
+            $service->destroy($ids);
+            Timoneiro::pushNotification("Successfully deleted {$displayName}");
+        } catch (\Exception $e) {
+            Timoneiro::pushNotification("Sorry it appears there was a problem deleting this {$displayName}", null, 'error');
+        }
 
         return redirect()->route("timoneiro.{$dataType->slug}.index");
     }