+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### added
+### Added
+- Title filed in general settings section
+- Toast notifications `Timoneiro::pushNotifications`
+- Delete functionality
 - package auto discovery
 
 ## [0.0.1-alpha.1] - 2019-09-17